# backend/docs/social-api-draft.openapi.yaml
...
openapi: 3.0.3
info:
  title: STROITECH Social API (draft)
  version: "0.1.0"
  description: >
    Планируемые маршруты. Префикс API в проде обычно /api (как у текущих /api/users, /api/feed).
servers:
  - url: /api
...
paths:
...
  /moderation/reports:
    post:
      tags: [moderation]
      summary: Пожаловаться на пост или комментарий ленты
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [targetType, targetId]
              properties:
                targetType: { type: string, enum: [feed_post, feed_comment] }
                targetId: { type: string, format: uuid }
                details: { type: string, maxLength: 2000 }
      responses:
        "200":
          description: OK
        "409":
          description: Уже есть жалоба от этого пользователя на объект
    get:
      tags: [moderation]
      summary: Очередь жалоб (только moderator)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, closed, all], default: pending }
        - $ref: "#/components/parameters/page"
        - name: limit
          in: query
          schema: { type: integer, default: 30, maximum: 100 }
      responses:
        "200":
          description: OK
  /moderation/reports/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [moderation]
      summary: Закрыть жалобу (только moderator)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [closed] }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...